backend/requirements.txt
# 后端依赖
fastapi[standard]
uvicorn[standard]
feedparser
httpx
apscheduler
anthropic
python-dotenv
